:lipstick: Bulma for AddAdministrator
Updated NavBar; Dashboard; AddAdministrator
