+++ packages/frontend/src/components/TheNavBar.vue
@@ -1,23 +1,20 @@
 <template>
-  <nav class="navbar" role="navigation" aria-label="main navigation">
-    <div class="navbar-brand">
-      <div class="navbar-item has-text-weight-bold">
-        <img src="../assets/FMDB_logo.png" height="auto">
-      </div>
-    </div>
-
-    <div id="navbarBasicExample" class="navbar-menu">
-      <div class="navbar-end">
-        <div class="navbar-item">
-          <div class="buttons">
-            <a class="button is-primary" @click="onClick">
-              <strong>Logout</strong>
-            </a>
-          </div>
+  <b-navbar type="is-dark">
+    <template slot="brand">
+      <b-navbar-item >
+        <span class="has-text-weight-bold">Fancy Moodle Discord Bot</span>
+      </b-navbar-item>
+    </template>
+    <template slot="end">
+      <b-navbar-item tag="div">
+        <div class="buttons">
+          <a class="button is-primary" @click="onClick">
+            <strong>Logout</strong>
+          </a>
         </div>
-      </div>
-    </div>
-  </nav>
+      </b-navbar-item>
+    </template>
+  </b-navbar>
 </template>
 
 <script>
